backend/requirements.txt
flask
flask-cors
flask-socketio
google-genai
python-dotenv
pymongo
flask-jwt-extended
supabase
